fix: consolidate pickBestFlight into shared lib, fix prefer-const lint error
- Export pickBestFlight from flight-status.ts with the robust scored implementation
- Import in live/route.ts, removing the duplicate
- Fix let→const for flights variable (ESLint prefer-const error)

Addresses Gemini code review finding (duplicated pickBestFlight) and ESLint CI failure.
